Special handling for "no user object for principal identifier" edge case
Throw a custom exception when the adapter cannot find a user object for the authenticated principal.  A special core.exception listener will intercept any such exception before Symfony and unauthenticate the user both locally and on the remote CAS server.  This does not solve the larger problem of CAS having authenticated a user that doesn't exist on our local system, but it's a much more graceful way to handle this for the user than recurring exceptions and/or 500 errors.